+++ CHANGELOG.md
@@ -12,6 +12,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Added `update_agent_status_by_hostname` public function to `Node` table. This
   function allows updating an agent’s status using the node’s hostname and the
   agent’s ID.
+- Added triage functionality for scoring with attributes of each raw event.
+  Previously, all detection events always returned a score of 0.0. It now
+  compares the values against the provided attribute correctly and returns a
+  score.
 
 ### Changed
 
@@ -22,6 +26,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   instead of Display trait.
 - `Filter::host_name` has been renamed to `Filter::hostname` to align with the
   naming convention used in other parts of the codebase.
+- Modified the `ValueKind` enum to support different types of input for packet
+  attribute triage.
 
 ### Fixed